requirements.txt
anyio==3.3.2
argon2-cffi==21.1.0
arrow==1.2.0
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
coverage==6.1.1
cryptography==35.0.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.9.0
django-grappelli==2.15.3
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
entrypoints==0.3
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
html5lib==1.1
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
json5==0.9.6
jsonschema==4.0.1
jupyter-client==7.0.6
jupyter-core==4.8.1
jupyter-server==1.11.1
jupyterlab==3.1.18
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclassic==0.3.2
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
twilio==7.0.0
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
